New python algorithms (7-30-16)
Ah yes, friends. calc 2 is finally over and I have more time to bust out
legit codez!!

The comparison module includes algorithms for finding the middle value
of three values, finding the largest and second second largest value in
a sequence, and finding the index of the last occurence of the largest
value in a sequence.  I also threw in an iterative algorithm for
reversing a sequence.

These are pretty simple; I haven't coded in a while cause calc 2 was
BROOTAL.

More to come. -JR